소켓 통신 막히던 부분 수정
원인
-  ReceiveOverlapped를 미리 걸어 놔야 클라이언트에서 보낸 데이터를 받을 준비를 한다는 소리임
- 커널함수 GetQueuedCompletionStatusEx에서 올라오는 lpCompletionKey를 재대로 사용하지 못했음

+++ SampleServer/CommonLib/TcpSocket.cpp
@@ -1,6 +1,6 @@
 #include "TcpSocket.h"
 
-TcpSocket::TcpSocket(const string& address, const uint8_t& port)
+TcpSocket::TcpSocket(const string& address, const int& port)
 	: Socket(address, port)
 {
 	_socketHandle = WSASocket(AF_INET, SOCK_STREAM, 0, NULL, 0, WSA_FLAG_OVERLAPPED);	// windows
